+++ features/step_definitions/parser.rb
@@ -10,6 +10,30 @@ Given /^the following specdown example file containing multiple executable codeb
   @readme = File.read "features/fixtures/multiple_codeblocks_per_section_example.markdown"
 end
 
+Given /^the following specdown example file containing both codeblocks and implicit specs:$/ do |string|
+  @readme = File.read "features/fixtures/codeblocks_and_implicits.markdown"
+end
+
+Given /^the following README:$/ do |string|
+  eval string
+end
+
+Then /^the root node should include the explicit code:$/ do |string|
+  eval string
+end
+
+Then /^the root node should include the undefined implicit specs:$/ do |string|
+  eval string
+end
+
+Given /^the following implicit specs:$/ do |string|
+  eval string
+end
+
+Then /^the code block and the implicit spec should be joined together:$/ do |string|
+  eval string
+end
+
 When /^I parse it into a tree:$/ do |string|
   eval string
 end