+++ lib/specdown/event_handlers/test_undefined.rb
@@ -1,0 +1,3 @@
+Specdown::EventServer.register :test_undefined do |test|
+  Specdown.reporter.print_undefined nil
+end