+++ lib/specdown/node.rb
@@ -1,11 +1,12 @@
 module Specdown
   class Node
-    attr_accessor :name, :code, :contents, :children, :parent
+    attr_accessor :name, :code, :undefined_implicits, :contents, :children, :parent
 
     def initialize
       @code = ''
       @contents = ''
       @children = []
+      @undefined_implicits = []
     end
 
     def code