+++ lib/specdown/pending.rb
@@ -1,0 +1,7 @@
+module Specdown
+  module Pending
+    def pending
+      raise Specdown::PendingException
+    end
+  end
+end