+++ lib/specdown/pending_exception.rb
@@ -1,0 +1,4 @@
+module Specdown
+  class PendingException < StandardError
+  end
+end