+++ lib/specdown/reporter.rb
@@ -40,6 +40,14 @@ module Specdown
       raise NotImplementedError
     end
 
+    def print_pending(test)
+      raise NotImplementedError
+    end
+
+    def print_undefined(test)
+      raise NotImplementedError
+    end
+
     def print_end
       raise NotImplementedError
     end