+++ lib/specdown/reporters/terminal_reporter.rb
@@ -10,6 +10,14 @@ module Specdown
       "F"
     end
 
+    def pending
+      "P"
+    end
+
+    def undefined
+      "U"
+    end
+
     def print_start
     end
 
@@ -37,8 +45,16 @@ module Specdown
       print success
     end
 
+    def print_pending(test)
+      print pending
+    end
+
     def print_failure(test)
       print failure
+    end
+
+    def print_undefined(test)
+      print undefined
     end
 
     def print_summary(runners)