+++ lib/specdown/runner/stats.rb
@@ -1,20 +1,31 @@
 module Specdown
   class Stats
-    attr_accessor :tests, :exceptions
+    attr_accessor :tests, :exceptions, :pending_exceptions, :undefined_implicits, :undefined_tests
     attr_reader :runner
 
     def initialize(source_runner=nil)
       @runner = source_runner
       @tests = 0
       @exceptions = []
+      @pending_exceptions = []
+      @undefined_implicits = []
+      @undefined_tests = 0
     end
 
     def successes
-      @tests - @exceptions.count
+      @tests - failures - pendings - undefined_tests
     end
 
     def failures
       @exceptions.count
+    end
+
+    def pendings
+      @pending_exceptions.count
+    end
+
+    def undefined
+      @undefined_implicits.count
     end
   end
 end