+++ lib/specdown/sandbox_decorators/pending.rb
@@ -1,0 +1,3 @@
+Specdown::SandboxFactory.decorate do |sandbox|
+  sandbox.extend Specdown::Pending 
+end